move interpolate_secret_key() from key-export-common to key-export-client
