+++ key-export-common/Cargo.toml
@@ -15,5 +15,4 @@ generic-ec = { git = "https://github.com/dfns-labs/generic-ec", branch = "m", de
 generic-ec-zkp = { git = "https://github.com/dfns-labs/generic-ec", branch = "m", default-features = false, features = ["serde", "alloc"] }
 
 [dev-dependencies]
-dfns-key-import-common = { path = "../key-import-common" }
 rand_dev = "0.1"
